services/gomini-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
grpcio==1.60.0
grpcio-tools==1.60.0
python-socketio==5.10.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
python-dotenv==1.0.0
cryptography>=41.0.0
pyjwt==2.8.0